requirements.txt
streamlit
pandas
scikit-learn
LabelEncoder
sklearn.preprocessing
pickle
scipy
matplotlip